pyproject.toml
[tool.poetry]
name = "elyza_tutorial"
version = "0.1.0"
description = ""
authors = ["tsomeya <tsomeya@student.42tokyo.jp>"]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.8.1"
matplotlib = "^3.4.1"
numpy = "^1.20.2"
mecab-python3 = "^1.0.3"
scikit-learn = "^0.24.1"
nltk = "^3.6.2"
seaborn = "^0.11.1"
pandas = "^1.2.4"
requests = "^2.25.1"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.8.2"
flake8 = "^3.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
